# table-serve-backend/openapi.yaml
openapi: 3.1.0
info:
  title: Table Serve API
  description: |
    REST API for the Table Serve platform - an NFC-based restaurant menu and ordering system.

    Three access levels:
    - **Super Admin** - Platform administrators (Table Serve company). Requires master password to register.
    - **Admin** - Restaurant owners and managers managing their menu, tables, and analytics.
    - **Customer** - Public-facing API for NFC-scanned menu viewing and order placement.
  version: 1.0.0
  contact:
    name: Table Serve Support
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.tableserve.app
    description: Production
...
paths:
...
  /api/customer/table/{token}:
    get:
      tags: [Customer]
      summary: Resolve NFC table token to restaurant and table info
      operationId: resolveTableToken
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Table and organization info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableResolveResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/customer/menu/{organizationId}:
    get:
      tags: [Customer]
      summary: Get menu for a restaurant
      operationId: getCustomerMenu
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Menu categories and items
        '404':
          $ref: '#/components/responses/NotFound'
  /api/customer/orders:
    post:
      tags: [Customer]
      summary: Place an order
      operationId: placeOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order placed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  orderId:
                    type: string
                  totalAmount:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Restaurant not accepting orders
        '404':
          $ref: '#/components/responses/NotFound'
  /api/customer/orders/{id}:
    get:
      tags: [Customer]
      summary: Get order status
      operationId: getOrderStatus
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Order details and status
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    CreateOrderRequest:
      type: object
      required: [tableToken, items]
      properties:
        tableToken:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        notes:
          type: string
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [menuItemId, quantity]
            properties:
              menuItemId:
                type: string
              quantity:
                type: integer
                minimum: 1
              notes:
                type: string
    TableResolveResponse:
      type: object
      properties:
        table:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            organizationId:
              type: string
        organization:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            primaryColor:
              type: string
            accentColor:
              type: string
            welcomeMessage:
              type: string
...